+++ src/modules/rooms/rooms.controller.ts
@@ -6,6 +6,7 @@ import {
   Param,
   ParseUUIDPipe,
   Post,
+  Query,
   Req,
   UploadedFile,
   UseGuards,
@@ -13,19 +14,20 @@ import {
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { AuthGuard } from '@nestjs/passport';
-import type { Request, Express } from 'express';
+import type { Request } from 'express';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { memoryStorage } from 'multer';
 
 import { routesV1 } from '@src/config';
 import { RoomsService } from './rooms.service';
 import { CreateRoomDto } from './dto/create-room.dto';
-import { CreatedRoomDto, RoomDetailsDto } from './dto/responses';
-import { CreateRoomDocs, GetRoomDocs } from './swagger-docs';
-import { Public } from '@src/common/decorators';
+import { CreatedRoomDto, PaginatedRoomsDto, RoomDetailsDto } from './dto/responses';
+import { CreateRoomDocs, GetRoomDocs, GetRoomsDocs } from './swagger-docs';
+import { GetRoomsQueryDto } from './dto/get-rooms.query.dto';
+import { RequirePermissions } from '@src/common/decorators';
+import { Permission } from '@prisma/client';
 
 @ApiTags(routesV1.rooms.root)
-@Public()
 @Controller(routesV1.version)
 export class RoomsController {
   constructor(private readonly roomsService: RoomsService) {}
@@ -59,13 +61,21 @@ export class RoomsController {
   }
 
   @Get(routesV1.rooms.findOne)
-  @UseGuards(AuthGuard('jwt'))
+  @RequirePermissions([Permission.USER_READ])
   @GetRoomDocs()
   findOne(
     @Req() req: Request,
     @Param('id', new ParseUUIDPipe()) id: string,
   ): Promise<RoomDetailsDto> {
-    const user = req.user as { id?: string } | undefined;
-    return this.roomsService.findOne(user?.id, id);
+    const user = req.user as { id: string };
+    return this.roomsService.findOne(user.id, id);
+  }
+
+  @Get(routesV1.rooms.findAll)
+  @RequirePermissions([Permission.USER_READ])
+  @GetRoomsDocs()
+  findAll(@Req() req: Request, @Query() query: GetRoomsQueryDto): Promise<PaginatedRoomsDto> {
+    const user = req.user as { id: string };
+    return this.roomsService.findAll(user?.id, query);
   }
 }